Move getFinancialAccountNumbers to importTransactions.ts
